+++ src/sfinance/__init__.py
@@ -2,5 +2,5 @@
 
 __author__ = """shiva kharbanda"""
 __email__ = 'shivakharbanda1311@gmail.com'
-__version__ = '0.1.0'
+__version__ = '0.2.0'